+++ src/App/App.js
@@ -1,18 +1,18 @@
-import logo from '../images/logo.svg';
-import '../css/App.css';
+import logo from '../assets/logo.svg';
+import style from './App.module.scss';
 
-import Group from './components/Group';
+import Group from './Group/Group';
 
 function App() {
   return (
-    <div className="App">
-      <header className="App-header">
-        <img src={logo} className="App-logo" alt="logo" />
+    <div className={style.App}>
+      <header className={style.header}>
+        <img src={logo} className={style.logo} alt="logo" />
         <p>
           Edit <code>src/App.js</code> and save to reload.
         </p>
         <a
-          className="App-link"
+          className={style.link}
           href="https://reactjs.org"
           target="_blank"
           rel="noopener noreferrer"
