+++ src/App/Group/Group.js
@@ -1,12 +1,12 @@
 // Brief Discription of out Group
-import '../../css/components/Group.css';
+import style from './Group.module.scss';
 
 function Group() {
   return (
-    <div className="Group">
+    <div>
       <h3>Group Brian</h3>
       <h4>Members</h4>
-      <ul className="Group-list">
+      <ul className={style.list}>
         <li>Brian</li>
         <li>Nancy</li>
         <li>Japheth</li>
