+++ src/App/Group/Group.module.scss
@@ -1,4 +1,4 @@
-.Group-list {
+.list {
   padding: 0;
   list-style-type: none;
 }
