+++ src/index.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import './css/index.css';
+import './index.module.scss';
 import App from './App/App';
 import reportWebVitals from './reportWebVitals';
